chore(TextInput): add tests for label and autocomplete
